Fix server crash caused by default room disconnect
When a user loads the home page, they are connected to a socket called
"default". If another user creates a room called "default", it will
crash the server if the user already on the home page refreshes. By
checking that the room exists before running most of the disconnect
code, this prevents an undefined property error.